requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
google-cloud-firestore>=2.15.0
langchain>=0.2.0
langchain-google-vertexai>=1.0.0
langchain-community>=0.2.0
wikipedia>=1.4.0
duckduckgo-search>=5.3.0
arxiv>=2.1.0
python-dotenv>=1.0.1
